Move Showcase/Wrapper to storybook components, fix key
