+++ .storybook/config.js
@@ -6,28 +6,9 @@ import styled from 'styled-components';
 
 import '../src/global-styles';
 import { COLORS } from '../src/constants';
+import Wrapper from './components/Wrapper';
 
-const OuterWrapper = styled.div`
-  position: absolute;
-  top: 0;
-  left: 0;
-  right: 0;
-  bottom: 0;
-`;
-const InnerWrapper = styled.div`
-  background: ${COLORS.white};
-  box-shadow: 1px 1px 2px rgba(0, 0, 0, 0.1);
-  margin: 2rem;
-  padding: 1.5rem;
-  max-width: 1100px;
-  border-radius: 4px;
-`;
-
-const WrapperDecorator = storyFn => (
-  <OuterWrapper>
-    <InnerWrapper>{storyFn()}</InnerWrapper>
-  </OuterWrapper>
-);
+const WrapperDecorator = storyFn => <Wrapper>{storyFn()}</Wrapper>;
 
 addDecorator(WrapperDecorator);
